+++ app/src/main/java/com/shuyun/qapp/bean/MinePrize.java
@@ -427,7 +427,7 @@ public class MinePrize implements Parcelable {
         private String mainImage;
         private String longImage;
         private String shortImage;
-        private int id;
+        private Long id;
         private double worthLower;
         private double worthUpper;
 
@@ -443,7 +443,7 @@ public class MinePrize implements Parcelable {
             mainImage = in.readString();
             longImage = in.readString();
             shortImage = in.readString();
-            id = in.readInt();
+            id = in.readLong();
             worthLower = in.readDouble();
             worthUpper = in.readDouble();
         }
@@ -516,11 +516,11 @@ public class MinePrize implements Parcelable {
             this.shortImage = shortImage;
         }
 
-        public int getId() {
+        public Long getId() {
             return id;
         }
 
-        public void setId(int id) {
+        public void setId(Long id) {
             this.id = id;
         }
 
@@ -554,7 +554,7 @@ public class MinePrize implements Parcelable {
             dest.writeString(mainImage);
             dest.writeString(longImage);
             dest.writeString(shortImage);
-            dest.writeInt(id);
+            dest.writeLong(id);
             dest.writeDouble(worthLower);
             dest.writeDouble(worthUpper);
         }
